+++ blocksuite/affine/all/package.json
@@ -146,13 +146,18 @@
     "./inlines/link": "./src/inlines/link/index.ts",
     "./inlines/link/store": "./src/inlines/link/store.ts",
     "./inlines/link/view": "./src/inlines/link/view.ts",
-    "./inlines/reference": "./src/inlines/reference.ts",
+    "./inlines/reference": "./src/inlines/reference/index.ts",
+    "./inlines/reference/store": "./src/inlines/reference/store.ts",
+    "./inlines/reference/view": "./src/inlines/reference/view.ts",
     "./inlines/preset": "./src/inlines/preset.ts",
     "./inlines/footnote": "./src/inlines/footnote/index.ts",
     "./inlines/footnote/view": "./src/inlines/footnote/view.ts",
     "./inlines/footnote/store": "./src/inlines/footnote/store.ts",
-    "./inlines/latex": "./src/inlines/latex.ts",
-    "./inlines/mention": "./src/inlines/mention.ts",
+    "./inlines/latex": "./src/inlines/latex/index.ts",
+    "./inlines/latex/store": "./src/inlines/latex/store.ts",
+    "./inlines/latex/view": "./src/inlines/latex/view.ts",
+    "./inlines/mention": "./src/inlines/mention/index.ts",
+    "./inlines/mention/view": "./src/inlines/mention/view.ts",
     "./widgets/drag-handle": "./src/widgets/drag-handle.ts",
     "./widgets/edgeless-auto-connect": "./src/widgets/edgeless-auto-connect.ts",
     "./widgets/edgeless-toolbar": "./src/widgets/edgeless-toolbar.ts",
@@ -215,7 +220,6 @@
     "./schemas": "./src/schemas.ts",
     "./model": "./src/model/index.ts",
     "./sync": "./src/sync/index.ts",
-    "./adapters": "./src/adapters/index.ts",
     "./extensions/store": "./src/extensions/store.ts",
     "./extensions/view": "./src/extensions/view.ts"
   },
