+++ blocksuite/affine/all/src/inlines/latex/store.ts
@@ -1,0 +1,1 @@
+export * from '@blocksuite/affine-inline-latex/store';
