+++ blocksuite/affine/all/src/inlines/mention/index.ts
