+++ blocksuite/affine/all/src/inlines/mention/view.ts
@@ -1,0 +1,1 @@
+export * from '@blocksuite/affine-inline-mention/view';
