+++ blocksuite/affine/all/src/inlines/reference/store.ts
@@ -1,0 +1,1 @@
+export * from '@blocksuite/affine-inline-reference/store';
