+++ blocksuite/affine/all/src/inlines/reference/view.ts
@@ -1,0 +1,1 @@
+export * from '@blocksuite/affine-inline-reference/view';
