+++ blocksuite/affine/inlines/latex/tsconfig.json
@@ -8,6 +8,7 @@
   "include": ["./src"],
   "references": [
     { "path": "../../components" },
+    { "path": "../../ext-loader" },
     { "path": "../reference" },
     { "path": "../../model" },
     { "path": "../../rich-text" },
