+++ blocksuite/affine/inlines/mention/src/view.ts
@@ -1,0 +1,21 @@
+import {
+  type ViewExtensionContext,
+  ViewExtensionProvider,
+} from '@blocksuite/affine-ext-loader';
+
+import { effects } from './effects';
+import { MentionInlineSpecExtension } from './inline-spec';
+
+export class MentionViewExtension extends ViewExtensionProvider {
+  override name = 'affine-mention-inline';
+
+  override effect(): void {
+    super.effect();
+    effects();
+  }
+
+  override setup(context: ViewExtensionContext): void {
+    super.setup(context);
+    context.register(MentionInlineSpecExtension);
+  }
+}
