--- blocksuite/affine/inlines/preset/src/adapters/plain-text/inline-delta.ts
@@ -1,8 +1,0 @@
-import { latexDeltaMarkdownAdapterMatch } from '@blocksuite/affine-inline-latex';
-import { referenceDeltaMarkdownAdapterMatch } from '@blocksuite/affine-inline-reference';
-import type { ExtensionType } from '@blocksuite/store';
-
-export const InlineDeltaToPlainTextAdapterExtensions: ExtensionType[] = [
-  referenceDeltaMarkdownAdapterMatch,
-  latexDeltaMarkdownAdapterMatch,
-];
