+++ blocksuite/affine/inlines/preset/src/inline-spec.ts
@@ -1,6 +1,3 @@
-import { inlineLatexExtensions } from '@blocksuite/affine-inline-latex';
-import { inlineMentionExtensions } from '@blocksuite/affine-inline-mention';
-import { inlineReferenceExtensions } from '@blocksuite/affine-inline-reference';
 import type { AffineTextAttributes } from '@blocksuite/affine-shared/types';
 import {
   type InlineRootElement,
@@ -104,7 +101,4 @@ export const InlineSpecExtensions: ExtensionType[] = [
   CodeInlineSpecExtension,
   BackgroundInlineSpecExtension,
   ColorInlineSpecExtension,
-  ...inlineReferenceExtensions,
-  ...inlineLatexExtensions,
-  ...inlineMentionExtensions,
 ];
